Rename "lynxkite.core" module to "lynxkite_core". Rename "lynxkite" package to "lynxkite-app".
