+++ docs/reference/lynxkite-core/workspace.md
@@ -1,1 +1,1 @@
-::: lynxkite.core.workspace
+::: lynxkite_core.workspace
