+++ examples/make_image_table.py
@@ -1,4 +1,4 @@
-from lynxkite.core.ops import op
+from lynxkite_core.ops import op
 import pandas as pd
 import base64
