+++ lynxkite-app/src/lynxkite_app/crdt.py
@@ -10,7 +10,7 @@ import pycrdt.websocket
 import pycrdt.store.file
 import uvicorn.protocols.utils
 import builtins
-from lynxkite.core import workspace, ops
+from lynxkite_core import workspace, ops
 
 router = fastapi.APIRouter()
